+++ pkg/deploy/taskdir/definitions/def_0_3.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"path"
 
+	"github.com/airplanedev/lib/pkg/api"
 	"github.com/airplanedev/lib/pkg/build"
-	"github.com/airplanedev/lib/pkg/deploy/api"
 	"github.com/goccy/go-yaml"
 	"github.com/pkg/errors"
 	"github.com/xeipuuv/gojsonschema"
