+++ pkg/deploy/taskdir/definitions/definitions.go
@@ -8,8 +8,8 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/airplanedev/lib/pkg/api"
 	"github.com/airplanedev/lib/pkg/build"
-	"github.com/airplanedev/lib/pkg/deploy/api"
 	"github.com/airplanedev/lib/pkg/utils/pathcase"
 	"github.com/mitchellh/mapstructure"
 	"github.com/pkg/errors"
