+++ pkg/deploy/taskdir/definitions/interface.go
@@ -3,8 +3,8 @@ package definitions
 import (
 	"context"
 
+	"github.com/airplanedev/lib/pkg/api"
 	"github.com/airplanedev/lib/pkg/build"
-	"github.com/airplanedev/lib/pkg/deploy/api"
 )
 
 type DefinitionInterface interface {
